refactor: create helpers for template & nodemailer
